Covert ExprTee from struct to class
TODO: complete node types
