--- ExprTree/ExprTree.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30002.166
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ExprTree", "ExprTree\ExprTree.vcxproj", "{2088F4C9-EFA5-46EA-A566-0F7BC346F84B}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2088F4C9-EFA5-46EA-A566-0F7BC346F84B}.Debug|x64.ActiveCfg = Debug|x64
-		{2088F4C9-EFA5-46EA-A566-0F7BC346F84B}.Debug|x64.Build.0 = Debug|x64
-		{2088F4C9-EFA5-46EA-A566-0F7BC346F84B}.Debug|x86.ActiveCfg = Debug|Win32
-		{2088F4C9-EFA5-46EA-A566-0F7BC346F84B}.Debug|x86.Build.0 = Debug|Win32
-		{2088F4C9-EFA5-46EA-A566-0F7BC346F84B}.Release|x64.ActiveCfg = Release|x64
-		{2088F4C9-EFA5-46EA-A566-0F7BC346F84B}.Release|x64.Build.0 = Release|x64
-		{2088F4C9-EFA5-46EA-A566-0F7BC346F84B}.Release|x86.ActiveCfg = Release|Win32
-		{2088F4C9-EFA5-46EA-A566-0F7BC346F84B}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {700DA116-CB9D-44D5-A9C4-6F6A2D0C845F}
-	EndGlobalSection
-EndGlobal
